Move firebase hooks logic to utils/effectBases.js
