+++ src/hooks/children.js
@@ -1,5 +1,5 @@
-import { useState, useEffect } from 'react';
-import { toData } from '../utils/helpers';
+import { useState, useEffect, useMemo } from 'react';
+import { getDataEffectBase, onSnapshotDataEffectBase } from '../utils/effectBases';
 
 /* ======================
  * === Children Hooks ===
@@ -31,7 +31,8 @@ export const useParentsChildren = accounts => {
   const [childRefs, setChildRefs] = useState([]);
   const [children, setChildren] = useState([]);
   const [loading, setLoading] = useState(true);
-  useEffect(parentChildRefsEffect(accounts, setChildRefs), [accounts]);
+  const parentRef = useMemo(() => accounts.parents.ref, [accounts]);
+  useEffect(parentChildRefsEffect(parentRef, setChildRefs), [parentRef]);
   useEffect(childDataEffect(childRefs, setChildren, setLoading), [childRefs]);
   return [children, loading];
 };
@@ -42,7 +43,8 @@ export const useParentsChildren = accounts => {
 export const useParentsLiveChildren = accounts => {
   const [childRefs, setChildRefs] = useState([]);
   const [children, setChildren] = useState([]);
-  useEffect(parentChildRefsEffect(accounts, setChildRefs), [accounts]);
+  const parentRef = useMemo(() => accounts.parents.ref, [accounts]);
+  useEffect(parentChildRefsEffect(parentRef, setChildRefs), [parentRef]);
   useEffect(liveChildDataEffect(childRefs, setChildren), [childRefs]);
   return children;
 };
@@ -54,55 +56,14 @@ export const useParentsLiveChildren = accounts => {
 /**
  * Fetch the data of each of the parent's children once.
  */
-const childDataEffect = (childRefs, setChildren, setLoading) => () => {
-  /**
-   * This is how you are supposed to run async code in an effect.
-   * See this github issue for details:
-   * https://github.com/facebook/react/issues/14326#issuecomment-441680293
-   */
-  async function run() {
-    setLoading(true);
-    setChildren(await getChildren(childRefs));
-    setLoading(false);
-  }
-  run();
-};
+const childDataEffect = getDataEffectBase(false);
 
 /**
- * Subscrive to each child's data.
+ * Subscribe to each child's data.
  */
-const liveChildDataEffect = (childRefs, setChildren) => {
-  let subs = [];
-  return () => {
-    const childrenMap = {};
-    childRefs.forEach(ref => {
-      const sub = ref.onSnapshot(childDoc => {
-        childrenMap[childDoc.id] = toData(childDoc);
-        setChildren(Object.values(childrenMap));
-      });
-      subs.push(sub);
-    });
-    // Cleanup subscriptions.
-    return () => {
-      subs.forEach(sub => sub());
-      subs = [];
-    };
-  };
-};
+const liveChildDataEffect = onSnapshotDataEffectBase(false);
 
 /**
  * Subscribe to the parents children references.
  */
-const parentChildRefsEffect = (accounts, setChildRefs) => () =>
-  accounts.parents.ref.onSnapshot(async parentDoc => {
-    setChildRefs(parentDoc.data().children || []);
-  });
-
-/* ====================
- * === Hook Helpers ===
- * ==================== */
-
-/**
- * Get the data of a list of child references.
- */
-const getChildren = refs => Promise.all(refs.map(async childRef => toData(await childRef.get())));
+const parentChildRefsEffect = onSnapshotDataEffectBase(true, doc => doc.data().children || []);
